use external-item decorator
